refactor: move execution implementation to new file
